Dictionary of Tuning Spaces

A simple mlr3misc::Dictionary storing objects of class TuningSpace. Each tuning space has an associated help page, see mlr_tuning_spaces_[id].

Format

R6::R6Class object inheriting from mlr3misc::Dictionary.

Methods

See mlr3misc::Dictionary.

S3 methods

as.data.table(dict, ..., objects = FALSE)
mlr3misc::Dictionary -> data.table::data.table()
Returns a data.table::data.table() with fields "key", "label", "learner", and "n_values" as columns. If objects is set to TRUE, the constructed objects are returned in the list column named object.

Examples

as.data.table(mlr_tuning_spaces)
#>                         key
#>  1:  classif.glmnet.default
#>  2:     classif.glmnet.rbv2
#>  3:    classif.kknn.default
#>  4:       classif.kknn.rbv2
#>  5:  classif.ranger.default
#>  6:     classif.ranger.rbv2
#>  7:   classif.rpart.default
#>  8:      classif.rpart.rbv2
#>  9:     classif.svm.default
#> 10:        classif.svm.rbv2
#> 11: classif.xgboost.default
#> 12:    classif.xgboost.rbv2
#> 13:     regr.glmnet.default
#> 14:        regr.glmnet.rbv2
#> 15:       regr.kknn.default
#> 16:          regr.kknn.rbv2
#> 17:     regr.ranger.default
#> 18:        regr.ranger.rbv2
#> 19:      regr.rpart.default
#> 20:         regr.rpart.rbv2
#> 21:        regr.svm.default
#> 22:           regr.svm.rbv2
#> 23:    regr.xgboost.default
#> 24:       regr.xgboost.rbv2
#>                         key
#>                                                            label
#>  1:   Default GLM with Elastic Net Regularization Classification
#>  2: RandomBot GLM with Elastic Net Regularization Classification
#>  3:                    Default k-Nearest-Neighbor Classification
#>  4:                  RandomBot k-Nearest-Neighbor Classification
#>  5:                                Default Ranger Classification
#>  6:                              RandomBot Ranger Classification
#>  7:                                  Default Classification Tree
#>  8:                                RandomBot Classification Tree
#>  9:                Default Support Vector Machine Classification
#> 10:              RandomBot Support Vector Machine Classification
#> 11:             Default Extreme Gradient Boosting Classification
#> 12:           RandomBot Extreme Gradient Boosting Classification
#> 13:       Default GLM with Elastic Net Regularization Regression
#> 14:     RandomBot GLM with Elastic Net Regularization Regression
#> 15:                        Default k-Nearest-Neighbor Regression
#> 16:                      RandomBot k-Nearest-Neighbor Regression
#> 17:                                    Default Ranger Regression
#> 18:                                  RandomBot Ranger Regression
#> 19:                                      Default Regression Tree
#> 20:                                    RandomBot Regression Tree
#> 21:                    Default Support Vector Machine Regression
#> 22:                  RandomBot Support Vector Machine Regression
#> 23:                 Default Extreme Gradient Boosting Regression
#> 24:               RandomBot Extreme Gradient Boosting Regression
#>                                                            label
#>             learner n_values
#>  1:  classif.glmnet        2
#>  2:  classif.glmnet        2
#>  3:    classif.kknn        3
#>  4:    classif.kknn        1
#>  5:  classif.ranger        4
#>  6:  classif.ranger        8
#>  7:   classif.rpart        3
#>  8:   classif.rpart        4
#>  9:     classif.svm        4
#> 10:     classif.svm        5
#> 11: classif.xgboost        8
#> 12: classif.xgboost       13
#> 13:     regr.glmnet        2
#> 14:     regr.glmnet        2
#> 15:       regr.kknn        3
#> 16:       regr.kknn        1
#> 17:     regr.ranger        4
#> 18:     regr.ranger        7
#> 19:      regr.rpart        3
#> 20:      regr.rpart        4
#> 21:        regr.svm        4
#> 22:        regr.svm        5
#> 23:    regr.xgboost        8
#> 24:    regr.xgboost       13
#>             learner n_values
mlr_tuning_spaces$get("classif.ranger.default")
#> <TuningSpace:classif.ranger.default>: Default Ranger Classification
#>                 id lower upper      levels logscale
#> 1:      mtry.ratio   0.0     1                FALSE
#> 2:         replace    NA    NA  TRUE,FALSE    FALSE
#> 3: sample.fraction   0.1     1                FALSE
#> 4:       num.trees   1.0  2000                FALSE
lts("classif.ranger.default")
#> <TuningSpace:classif.ranger.default>: Default Ranger Classification
#>                 id lower upper      levels logscale
#> 1:      mtry.ratio   0.0     1                FALSE
#> 2:         replace    NA    NA  TRUE,FALSE    FALSE
#> 3: sample.fraction   0.1     1                FALSE
#> 4:       num.trees   1.0  2000                FALSE