Dictionary of Tuning Spaces

A simple mlr3misc::Dictionary storing objects of class TuningSpace. Each tuning space has an associated help page, see mlr_tuning_spaces_[id].

Format

R6::R6Class object inheriting from mlr3misc::Dictionary.

Methods

See mlr3misc::Dictionary.

S3 methods

as.data.table(dict, ..., objects = FALSE)
mlr3misc::Dictionary -> data.table::data.table()
Returns a data.table::data.table() with fields "key", "label", "learner", and "n_values" as columns. If objects is set to TRUE, the constructed objects are returned in the list column named object.

Examples

as.data.table(mlr_tuning_spaces)
#> Key: <key>
#>                         key                                 label
#>                      <char>                                <char>
#>  1:  classif.glmnet.default       Classification GLM with Default
#>  2:     classif.glmnet.rbv1     Classification GLM with RandomBot
#>  3:     classif.glmnet.rbv2     Classification GLM with RandomBot
#>  4:    classif.kknn.default      Classification KKNN with Default
#>  5:       classif.kknn.rbv1    Classification KKNN with RandomBot
#>  6:       classif.kknn.rbv2    Classification KKNN with RandomBot
#>  7:  classif.ranger.default    Classification Ranger with Default
#>  8:     classif.ranger.rbv1  Classification Ranger with RandomBot
#>  9:     classif.ranger.rbv2  Classification Ranger with RandomBot
#> 10:   classif.rpart.default     Classification Rpart with Default
#> 11:      classif.rpart.rbv1   Classification Rpart with RandomBot
#> 12:      classif.rpart.rbv2   Classification Rpart with RandomBot
#> 13:     classif.svm.default       Classification SVM with Default
#> 14:        classif.svm.rbv1     Classification SVM with RandomBot
#> 15:        classif.svm.rbv2     Classification SVM with RandomBot
#> 16: classif.xgboost.default   Classification XGBoost with Default
#> 17:    classif.xgboost.rbv1 Classification XGBoost with RandomBot
#> 18:    classif.xgboost.rbv2 Classification XGBoost with RandomBot
#> 19:     regr.glmnet.default           Regression GLM with Default
#> 20:        regr.glmnet.rbv1         Regression GLM with RandomBot
#> 21:        regr.glmnet.rbv2         Regression GLM with RandomBot
#> 22:       regr.kknn.default          Regression KKNN with Default
#> 23:          regr.kknn.rbv1        Regression KKNN with RandomBot
#> 24:          regr.kknn.rbv2        Regression KKNN with RandomBot
#> 25:     regr.ranger.default        Regression Ranger with Default
#> 26:        regr.ranger.rbv1      Regression Ranger with RandomBot
#> 27:        regr.ranger.rbv2      Regression Ranger with RandomBot
#> 28:      regr.rpart.default         Regression Rpart with Default
#> 29:         regr.rpart.rbv1       Regression Rpart with RandomBot
#> 30:         regr.rpart.rbv2       Regression Rpart with RandomBot
#> 31:        regr.svm.default           Regression SVM with Default
#> 32:           regr.svm.rbv1         Regression SVM with RandomBot
#> 33:           regr.svm.rbv2         Regression SVM with RandomBot
#> 34:    regr.xgboost.default       Regression XGBoost with Default
#> 35:       regr.xgboost.rbv1     Regression XGBoost with RandomBot
#> 36:       regr.xgboost.rbv2     Regression XGBoost with RandomBot
#>                         key                                 label
#>             learner n_values
#>              <char>    <int>
#>  1:  classif.glmnet        2
#>  2:  classif.glmnet        2
#>  3:  classif.glmnet        2
#>  4:    classif.kknn        3
#>  5:    classif.kknn        1
#>  6:    classif.kknn        1
#>  7:  classif.ranger        4
#>  8:  classif.ranger        6
#>  9:  classif.ranger        8
#> 10:   classif.rpart        3
#> 11:   classif.rpart        4
#> 12:   classif.rpart        4
#> 13:     classif.svm        4
#> 14:     classif.svm        4
#> 15:     classif.svm        5
#> 16: classif.xgboost        8
#> 17: classif.xgboost       10
#> 18: classif.xgboost       13
#> 19:     regr.glmnet        2
#> 20:     regr.glmnet        2
#> 21:     regr.glmnet        2
#> 22:       regr.kknn        3
#> 23:       regr.kknn        1
#> 24:       regr.kknn        1
#> 25:     regr.ranger        4
#> 26:     regr.ranger        6
#> 27:     regr.ranger        7
#> 28:      regr.rpart        3
#> 29:      regr.rpart        4
#> 30:      regr.rpart        4
#> 31:        regr.svm        4
#> 32:        regr.svm        4
#> 33:        regr.svm        5
#> 34:    regr.xgboost        8
#> 35:    regr.xgboost       10
#> 36:    regr.xgboost       13
#>             learner n_values
mlr_tuning_spaces$get("classif.ranger.default")
#> <TuningSpace:classif.ranger.default>: Classification Ranger with Default
#>                 id lower upper levels logscale
#>             <char> <num> <num> <list>   <lgcl>
#> 1:      mtry.ratio   0.0     1           FALSE
#> 2:         replace    NA    NA           FALSE
#> 3: sample.fraction   0.1     1           FALSE
#> 4:       num.trees   1.0  2000           FALSE
lts("classif.ranger.default")
#> <TuningSpace:classif.ranger.default>: Classification Ranger with Default
#>                 id lower upper levels logscale
#>             <char> <num> <num> <list>   <lgcl>
#> 1:      mtry.ratio   0.0     1           FALSE
#> 2:         replace    NA    NA           FALSE
#> 3: sample.fraction   0.1     1           FALSE
#> 4:       num.trees   1.0  2000           FALSE